lesson_2/expected_output/pyproject.toml
[project]
name = "lesson-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.46",
    "httpx>=0.27.0",
    "mcp>=1.26.0",
]
